Removed String based examples and new implementation untested ones

+++ examples/ESP8266/EchoBot/EchoBot.ino
@@ -18,8 +18,8 @@ char password[] = "YYYYYY"; // your network key
 WiFiClientSecure client;
 UniversalTelegramBot bot(BOTtoken, client);
 
-int Bot_mtbs = 1000; //mean time between scan messages
-long Bot_lasttime;   //last time messages' scan has been done
+unsigned long Bot_mtbs = 1000; //mean time between scan messages
+unsigned long Bot_lasttime;   //last time messages' scan has been done
 
 void setup() {
   Serial.begin(115200);
